fix(spanner): Refactoring: 1. Renamed files and variables for readability 2. Moved functions to more appropriate packages.
