+++ import_file/csv_data.go
@@ -1,4 +1,4 @@
-package import_data
+package import_file
 
 import (
 	"context"
