+++ import_file/csv_schema.go
@@ -1,4 +1,4 @@
-package import_data
+package import_file
 
 import (
 	"context"
